fix: e2e test
